4th day

+++ 02Nap/02SikidomokTerulete/IPlane.cs
@@ -4,6 +4,10 @@
     {
         //a public kolcsszóra nincs szükség, mert 
         //ami a felületen látszik az definíció szerint publikus
-        int Area();
+        double Area();
+
+        //ha például a felület kér egy ilyen tulajdonságot,
+        //ez közösen megoldható például egy abszrakt osztállyal
+        //string Name { get; set; }
     }
 }
